+++ packages/client/src/views/CorrectionForm/CorrectionSummary.tsx
@@ -48,7 +48,7 @@ import {
 } from '@client/forms'
 import { lookup } from 'country-data'
 import { ActionPageLight, TableView } from '@opencrvs/components/lib/interface'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 import {
   SuccessButton,
   SecondaryButton,
