+++ packages/client/src/views/PrintCertificate/IDVerifier.tsx
@@ -11,7 +11,7 @@
  */
 import * as React from 'react'
 import styled from '@client/styledComponents'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 import {
   SuccessButton,
   DangerButton,
