+++ packages/client/src/views/PrintCertificate/ReviewCertificateAction.tsx
@@ -16,7 +16,7 @@ import {
   TertiaryButton
 } from '@opencrvs/components/lib/buttons'
 import { Check } from '@opencrvs/components/lib/icons'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 
 import { ActionPageLight } from '@opencrvs/components/lib/interface'
 import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
