+++ packages/client/src/views/PrintCertificate/collectorForm/CollectorForm.tsx
@@ -11,7 +11,7 @@
  */
 import { PrimaryButton, TertiaryButton } from '@opencrvs/components/lib/buttons'
 import { ErrorText } from '@opencrvs/components/lib/ErrorText'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 
 import { ActionPageLight } from '@opencrvs/components/lib/interface'
 import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
