+++ packages/client/src/views/SysAdmin/Config/Certificates/Certificates.tsx
@@ -57,7 +57,7 @@ import {
   printCertificate,
   downloadFile
 } from '@client/views/PrintCertificate/PDFUtils'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 import { ListViewSimplified } from '@opencrvs/components/lib/interface/ListViewSimplified/ListViewSimplified'
 import { updateOfflineCertificate } from '@client/offline/actions'
 import { ICertificateTemplateData } from '@client/utils/referenceApi'
