+++ packages/client/src/views/SysAdmin/Config/Forms/Home/FormConfigHome.tsx
@@ -14,7 +14,7 @@ import * as React from 'react'
 import { useIntl } from 'react-intl'
 import { useSelector } from 'react-redux'
 import { SysAdminContentWrapper } from '@client/views/SysAdmin/SysAdminContentWrapper'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 import {
   messages,
   draftTabsMessages
