+++ packages/client/src/views/SysAdmin/Config/Forms/Wizard/FormConfigSettings.tsx
@@ -17,7 +17,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { ListViewItemSimplified } from '@opencrvs/components/lib/ListViewSimplified'
 import { Toast, NOTIFICATION_TYPE } from '@opencrvs/components/lib/Toast'
 import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 import { messages } from '@client/i18n/messages/views/formConfig'
 import { buttonMessages } from '@client/i18n/messages'
 import { EMPTY_STRING } from '@client/utils/constants'
