+++ packages/client/src/views/SysAdmin/Performance/WorkflowStatus.tsx
@@ -58,10 +58,7 @@ import subYears from 'date-fns/subYears'
 import differenceInSeconds from 'date-fns/differenceInSeconds'
 import { messages as statusMessages } from '@client/i18n/messages/views/registrarHome'
 import { colors } from '@opencrvs/components/lib/colors'
-import {
-  Content,
-  ContentSize
-} from '@opencrvs/components/lib/interface/Content'
+import { Content, ContentSize } from '@opencrvs/components/lib/Content'
 import { Spinner } from '@opencrvs/components/lib/Spinner'
 import { TableView } from '@opencrvs/components/lib/interface/TableView'
 import { Pagination } from '@opencrvs/components/lib/Pagination'
