+++ packages/client/src/views/SysAdmin/Team/user/userCreation/UserReviewForm.tsx
@@ -77,7 +77,7 @@ import {
   IListViewItemSimplifiedProps
 } from '@opencrvs/components/lib/interface/ListViewSimplified/ListViewSimplified'
 import styled from 'styled-components'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 
 export interface IUserReviewFormProps {
   userId?: string
