+++ packages/client/src/views/UserSetup/CreatePassword.tsx
@@ -24,7 +24,7 @@ import {
 } from '@client/components/ProtectedAccount'
 import { messages } from '@client/i18n/messages/views/userSetup'
 import { buttonMessages } from '@client/i18n/messages'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 import { EMPTY_STRING } from '@client/utils/constants'
 
 const GlobalError = styled.div`
