+++ packages/client/src/views/UserSetup/SecurityQuestionView.tsx
@@ -29,7 +29,7 @@ import {
   IProtectedAccountSetupData,
   ISecurityQuestionAnswer
 } from '@client/components/ProtectedAccount'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 
 const EMPTY_VALUE = ''
 const VISIBLE_QUESTION = 3
