+++ packages/client/src/views/UserSetup/SetupReviewPage.tsx
@@ -41,7 +41,7 @@ import {
 import { Check } from '@opencrvs/components/lib/icons'
 import { activateUserMutation } from '@client/views/UserSetup/queries'
 import { messages } from '@client/i18n/messages/views/userSetup'
-import { Content } from '@opencrvs/components/lib/interface/Content'
+import { Content } from '@opencrvs/components/lib/Content'
 import {
   SubmitActivateUserMutation,
   SubmitActivateUserMutationVariables
