+++ packages/components/src/Content/index.ts
@@ -9,4 +9,5 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
+export * from './Container'
 export * from './Content'
